+++ crates/kobold/src/event.rs
@@ -3,8 +3,11 @@
 use std::marker::PhantomData;
 use std::ops::Deref;
 
-use wasm_bindgen::JsCast;
+use wasm_bindgen::closure::Closure;
+use wasm_bindgen::{JsCast, JsValue};
 use web_sys::HtmlElement;
+
+use crate::{Element, Html, Mountable};
 
 /// Smart wrapper around a [`web_sys::Event`](web_sys::Event) which includes type
 /// information for the target element of said event.
@@ -29,6 +32,15 @@ where
     }
 }
 
+impl<T, E> From<web_sys::Event> for Event<T, E> {
+    fn from(event: web_sys::Event) -> Self {
+        Event {
+            event,
+            _target: PhantomData,
+        }
+    }
+}
+
 impl<T, E> Event<T, E> {
     pub fn target(&self) -> T
     where
@@ -50,8 +62,67 @@ impl<T, E> Event<T, E> {
     }
 }
 
-pub fn event_handler<T, E>(handler: impl Fn(Event<T, E>)) -> EventHandler<impl Fn(Event<T, E>)> {
-    EventHandler(handler)
+pub fn event_handler<E>(handler: impl Fn(E) + 'static) -> EventHandler<impl Fn(web_sys::Event) + 'static>
+where
+    E: From<web_sys::Event>,
+{
+    EventHandler(move |event| handler(E::from(event)))
 }
 
 pub struct EventHandler<F>(F);
+
+pub struct ClosureProduct<F> {
+    js: JsValue,
+    boxed: Box<F>,
+}
+
+impl<F> ClosureProduct<F>
+where
+    F: FnMut(web_sys::Event) + 'static,
+{
+    fn make(f: F) -> Self {
+        let raw = Box::into_raw(Box::new(f));
+
+        let js = Closure::wrap(unsafe { Box::from_raw(raw) } as Box<dyn FnMut(web_sys::Event)>)
+            .into_js_value();
+
+        // `into_js_value` will _forget_ the previous Box, so we can safely reconstruct it
+        let boxed = unsafe { Box::from_raw(raw) };
+
+        ClosureProduct { js, boxed }
+    }
+
+    fn update(&mut self, f: F) {
+        *self.boxed = f;
+    }
+}
+
+impl<F> Html for EventHandler<F>
+where
+    F: Fn(web_sys::Event) + 'static,
+{
+    type Product = ClosureProduct<F>;
+
+    fn build(self) -> Self::Product {
+        ClosureProduct::make(self.0)
+    }
+
+    fn update(self, p: &mut Self::Product) {
+        p.update(self.0)
+    }
+}
+
+impl<F> Mountable for ClosureProduct<F>
+where
+    F: 'static,
+{
+    type Js = JsValue;
+
+    fn el(&self) -> &Element {
+        panic!("Closure is not an element");
+    }
+
+    fn js(&self) -> &JsValue {
+        &self.js
+    }
+}
