+++ crates/kobold/src/lib.rs
@@ -285,7 +285,6 @@ pub use kobold_macros::html;
 
 use wasm_bindgen::{JsCast, JsValue};
 
-mod render_fn;
 mod value;
 
 pub mod attribute;
@@ -293,7 +292,6 @@ pub mod branching;
 pub mod dom;
 pub mod event;
 pub mod list;
-pub mod state;
 pub mod stateful;
 pub mod util;
 
@@ -302,8 +300,8 @@ pub mod prelude {
     pub use crate::{class, bind};
     pub use crate::event::{Event, KeyboardEvent, MouseEvent};
     pub use crate::list::ListIteratorExt as _;
-    pub use crate::state::{stateful, Hook, Then, Signal};
-    pub use crate::stateful::{IntoState, ShouldRender, WeakHook};
+    pub use crate::stateful::{stateful, Hook, Then, Signal, IntoState};
+    // pub use crate::stateful::{ShouldRender, WeakHook};
     // pub use crate::stateful::{stateful, Hook, IntoState, ShouldRender, WeakHook};
     pub use crate::value::{StrExt as _, Stringify as _};
     pub use crate::{component, html, Html};
